% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrp-package.R
\docType{package}
\name{burrp}
\alias{burrp}
\alias{burrp-package}
\title{Tools to Import and Process 'PortSwigger' 'Burp' Proxy Data}
\description{
Tools to Import and Process 'PortSwigger' 'Burp' Proxy Data
}
\author{
Bob Rudis (bob@rud.is)
}
