% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mime_is}
\alias{mime_is}
\alias{method_is}
\alias{status_is}
\alias{protocol_is}
\alias{extension_is}
\alias{port_is}
\title{Functional helpers}
\usage{
mime_is(x, mimetype)

method_is(x, method)

status_is(x, status)

protocol_is(x, protocol)

extension_is(x, extension)

port_is(x, port)
}
\arguments{
\item{x}{object}

\item{mimetype}{character}

\item{method}{character}

\item{status}{character}

\item{protocol}{character}

\item{extension}{character}

\item{port}{character}
}
\description{
Functional helpers
}
