% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{read_burp}
\alias{read_burp}
\title{Read in a Burp proxy XML export file}
\usage{
read_burp(burp_file)
}
\arguments{
\item{burp_file}{path to a Burp proxy XML export file}
}
\value{
a \code{tibble}
}
\description{
For now, this function expects the \code{request} and \code{response}
elements to be base 64 encoded.
}
\details{
Eventually the \code{request} and \code{response} objects will be
turned into \code{httr} compatible objects and we'll have
an \code{as_har()} function to turn the entire structure into a
\code{HARtools} object.
}
\examples{
library(dplyr)

system.file("extdata", "burp.xml", package="burrp") \%>\%
  read_burp() \%>\%
  glimpse()
}
