% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{read_burp}
\alias{read_burp}
\title{Read in a Burp proxy XML export file}
\usage{
read_burp(burp_file, convert_response = TRUE, convert_request = TRUE)
}
\arguments{
\item{burp_file}{path to a Burp proxy XML export file}

\item{convert_response}{if \code{TRUE}, turn the \code{response} record into
and \code{httr} \code{response} object. If the \code{response} record cannot
be turned into an \code{httr} \code{response} object a warning will
be issued and the raw \code{response} record will be returned.}

\item{convert_request}{if \code{TRUE}, turn the \code{request} record into
and \code{httr} \code{request} object. If the \code{request} record cannot
be turned into an \code{httr} \code{request} object a warning will
be issued and the raw \code{request} record will be returned.}
}
\value{
a \code{tibble}
}
\description{
For now, this function expects the \code{request} and \code{response}
elements to be base 64 encoded.
}
\details{
Eventually there will likely be an \code{as_har()} function to turn the
entire structure into a \code{HARtools} object.
}
\examples{
library(dplyr)

system.file("extdata", "hottest_year.xml", package="burrp") \%>\%
  read_burp() \%>\%
  glimpse()
}
